new revise project created

+++ Tracker/src/main/java/com/newVersion/example/entities/CoronaVirusData.java
@@ -3,13 +3,13 @@ package com.newVersion.example.entities;
 public class CoronaVirusData {
 
 	private String country;
-	private int total_Confirmed_Cases;
-	private int total_Deaths;
-	private int total_Recovered;
+	private double total_Confirmed_Cases;
+	private double total_Deaths;
+	private double total_Recovered;
 	
 
-	public CoronaVirusData(String country,int total_Confirmed_Cases, int total_Deaths,
-			int total_Recovered) {
+	public CoronaVirusData(String country,double total_Confirmed_Cases, double total_Deaths,
+			double total_Recovered) {
 		super();
 		this.country = country;
 		this.total_Confirmed_Cases = total_Confirmed_Cases;
@@ -25,24 +25,24 @@ public class CoronaVirusData {
 		this.country = country;
 	}
 
-	public void setTotal_Confirmed_Cases(int total_Confirmed_Cases) {
+	public void setTotal_Confirmed_Cases(double total_Confirmed_Cases) {
 		this.total_Confirmed_Cases = total_Confirmed_Cases;
 	}
-	public int getTotal_Confirmed_Cases() {
+	public double getTotal_Confirmed_Cases() {
 		return total_Confirmed_Cases;
 	}
 	
-	public void setTotal_Deaths(int total_Deaths) {
+	public void setTotal_Deaths(double total_Deaths) {
 		this.total_Deaths = total_Deaths;
 	}
-	public int getTotal_Deaths() {
+	public double getTotal_Deaths() {
 		return total_Deaths;
 	}
 	
-	public void setTotal_Recovered(int total_Recovered) {
+	public void setTotal_Recovered(double total_Recovered) {
 		this.total_Recovered = total_Recovered;
 	}
-	public int getTotal_Recovered() {
+	public double getTotal_Recovered() {
 		return total_Recovered;
 	}
